week two
Learning hours:8h
Learning hours:8h
I learned the required content of c language and did some practice .I have trouble dealing with some questions which require a high understanding of for ciculation. but I
will practice mor to master it .I also learned the basic information of LINUX.It is hard to master because I have never been exposed to it before but I will try my best
to understand it.